[ :wrench: | #64 | fix ] Coupon 엔티티 필드 수정
- 필드 타입 Integer -> Long 으로 수정

+++ src/main/java/com/objects/marketbridge/domain/model/Coupon.java
@@ -21,22 +21,22 @@ public class Coupon extends BaseEntity {
 
     private String name;
 
-    private Integer price;
+    private Long price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
     private Product product;
 
-    private Integer count;
+    private Long count;
 
-    private Integer minimumPrice;
+    private Long minimumPrice;
 
     private LocalDateTime startDate;
 
     private LocalDateTime endDate;
 
     @Builder
-    private Coupon(String name, Integer price, Product product, Integer count, Integer minimumPrice, LocalDateTime startDate, LocalDateTime endDate) {
+    private Coupon(String name, Long price, Product product, Long count, Long minimumPrice, LocalDateTime startDate, LocalDateTime endDate) {
         this.name = name;
         this.price = price;
         this.product = product;
